Add word2vec

+++ models/encode.py
@@ -81,7 +81,8 @@ class Encoder(object):
             print(f"Model does not exists, Downloading model: {self.model}")
             download_from_url(model_download_url, model_download_path)
             extract_file(model_download_path, model_path)
-            os.remove(model_download_path)
+            if os.path.exists(model_download_path):
+                os.remove(model_download_path)
             print(f"Model downloaded successfully!")
         return model_path
